% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talaan.R
\docType{package}
\name{talaan}
\alias{talaan-package}
\alias{talaan}
\title{An Interface to the Philippine Civil Registration and Vital Statistics Data}
\description{
The Philippine Statistics Authority (PSA) makes civil registration data
publicly available through their monthly updates on numbers of births,
deaths, and marriages released via their website. This package provides
utilities for accessing and processing the civil registration and vital
statistics data released by the PSA.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
